完成考试科目修改

+++ src/main/java/com/examw/netplatform/service/admin/settings/impl/SubjectServiceImpl.java
@@ -11,11 +11,13 @@ import org.springframework.beans.BeanUtils;
 
 import com.examw.model.DataGrid;
 import com.examw.netplatform.dao.admin.settings.AreaMapper;
+import com.examw.netplatform.dao.admin.settings.ChapterMapper;
 import com.examw.netplatform.dao.admin.settings.SubjectMapper;
 import com.examw.netplatform.domain.admin.settings.Area;
+import com.examw.netplatform.domain.admin.settings.Exam;
 import com.examw.netplatform.domain.admin.settings.Subject;
-import com.examw.netplatform.model.admin.settings.AreaInfo;
 import com.examw.netplatform.model.admin.settings.SubjectInfo;
+import com.examw.netplatform.service.admin.settings.IExamService;
 import com.examw.netplatform.service.admin.settings.ISubjectService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
@@ -28,6 +30,8 @@ public class SubjectServiceImpl implements ISubjectService {
 	private static final Logger logger = Logger.getLogger(SubjectServiceImpl.class);
 	private SubjectMapper subjectDao;
 	private AreaMapper areaDao;
+	private ChapterMapper chapterDao;
+	private IExamService examService;
 	private Map<Integer, String> statusMap;
 	/**
 	 * 设置科目数据接口。
@@ -46,6 +50,24 @@ public class SubjectServiceImpl implements ISubjectService {
 	public void setAreaDao(AreaMapper areaDao) {
 		logger.debug("注入地区数据接口...");
 		this.areaDao = areaDao;
+	}
+	/**
+	 * 设置章节数据接口。
+	 * @param chapterDao 
+	 *	  章节数据接口。
+	 */
+	public void setChapterDao(ChapterMapper chapterDao) {
+		logger.debug("注入章节数据接口...");
+		this.chapterDao = chapterDao;
+	}
+	/**
+	 * 设置考试服务接口。
+	 * @param examService 
+	 *	  考试服务接口。
+	 */
+	public void setExamService(IExamService examService) {
+		logger.debug("注入考试服务接口...");
+		this.examService = examService;
 	}
 	/**
 	 * 设置状态值名称集合。
@@ -83,7 +105,7 @@ public class SubjectServiceImpl implements ISubjectService {
 	public DataGrid<SubjectInfo> datagrid(SubjectInfo info) {
 		logger.debug("查询数据...");
 		//分页/排序
-		PageHelper.startPage(info.getPage(), info.getRows(), StringUtils.trimToEmpty(info.getOrder()) + " " + StringUtils.trimToEmpty(info.getSort()));
+		PageHelper.startPage(info.getPage(), info.getRows(), StringUtils.trimToEmpty(info.getSort()) + " " + StringUtils.trimToEmpty(info.getOrder()));
 		//查询数据
 		final List<Subject> list = this.subjectDao.findSubjects(info);
 		//分页信息
@@ -106,16 +128,34 @@ public class SubjectServiceImpl implements ISubjectService {
 		}
 		return list;
 	}
-	/*
-	 * 数据类型转换。
-	 * @see com.examw.netplatform.service.admin.settings.ISubjectService#conversion(com.examw.netplatform.domain.admin.settings.Subject)
-	 */
-	@Override
-	public SubjectInfo conversion(Subject subject) {
+	//数据类型转换
+	private SubjectInfo conversion(Subject data) {
 		logger.debug("数据类型Subject -> SubjectInfo ...");
-		SubjectInfo info = (SubjectInfo)subject;
-		info.setStatusName(this.loadStatusName(subject.getStatus()));
-		return info;
+		if(data != null){
+			final SubjectInfo info = new SubjectInfo();
+			BeanUtils.copyProperties(data, info);
+			info.setStatusName(this.loadStatusName(info.getStatus()));
+			//考试
+			if(StringUtils.isNotBlank(info.getExamId()) && StringUtils.isBlank(info.getExamName())){
+				final Exam exam = this.examService.loadExam(info.getExamId());
+				if(exam != null) info.setExamName(exam.getName());
+			}
+			//地区
+			final List<String> idList = new ArrayList<String>(), nameList = new ArrayList<String>();
+			final List<Area> areas = this.areaDao.findAreasBySubject(info.getId());
+			if(areas != null && areas.size() > 0){
+				for(Area area : areas){
+					if(area == null) continue;
+					idList.add(area.getId());
+					nameList.add(area.getName());
+				}
+			}
+			info.setAreaIds(idList.toArray(new String[0]));
+			info.setAreaNames(nameList.toArray(new String[0]));
+			//
+			return info;
+		}
+		return null;
 	}
 	/*
 	 * 加载考试科目数据。
@@ -127,22 +167,6 @@ public class SubjectServiceImpl implements ISubjectService {
 		return this.changeModel(this.subjectDao.findSubjectsByExam(examId));
 	}
 	/*
-	 * 加载科目地区。
-	 * @see com.examw.netplatform.service.admin.settings.ISubjectService#loadSubjectAreas(java.lang.String)
-	 */
-	@Override
-	public List<AreaInfo> loadSubjectAreas(String subjectId) {
-		logger.debug("加载科目["+subjectId+"]地区...");
-		final List<AreaInfo> list = new ArrayList<AreaInfo>();
-		final List<Area> areas = this.areaDao.findAreasBySubject(subjectId);
-		if(areas != null && areas.size() > 0){
-			for(Area area : areas){
-				list.add((AreaInfo)area);
-			}
-		}
-		return list;
-	}
-	/*
 	 * 更新科目。
 	 * @see com.examw.netplatform.service.admin.settings.ISubjectService#update(com.examw.netplatform.model.admin.settings.SubjectInfo)
 	 */
@@ -151,6 +175,11 @@ public class SubjectServiceImpl implements ISubjectService {
 		logger.debug("更新科目...");
 		if(info == null) return null;
 		boolean isAdded = false;
+		//检查考试
+		if(StringUtils.isBlank(info.getExamId()) || this.examService.loadExam(info.getExamId()) == null){
+			throw new RuntimeException("科目所属考试["+info.getExamId()+"]不存在!");
+		}
+		//
 		Subject data = StringUtils.isBlank(info.getId()) ? null : this.subjectDao.getSubject(info.getId());
 		if(isAdded = (data == null)){
 			if(StringUtils.isEmpty(info.getId()))
@@ -171,7 +200,17 @@ public class SubjectServiceImpl implements ISubjectService {
 		}else {
 			logger.debug("更新科目...");
 			this.subjectDao.updateSubject(data);
+			//删除科目地区
+			this.subjectDao.deleteSubjectAreas(data.getId());
 		}
+		//添加地区
+		if(data.getAreaIds() != null && data.getAreaIds().length > 0){
+			for(String areaId : data.getAreaIds()){
+				if(StringUtils.isBlank(areaId) || this.areaDao.getArea(areaId) == null) continue;
+				this.subjectDao.insertSubjectArea(data.getId(), areaId);
+			}
+		}
+		//
 		return this.conversion(data);
 	}
 	/*
@@ -183,7 +222,10 @@ public class SubjectServiceImpl implements ISubjectService {
 		logger.debug("删除科目..." + StringUtils.join(ids,","));
 		if(ids != null && ids.length > 0){
 			for(String id : ids){
-				if(StringUtils.isBlank(id)) continue;
+				if(StringUtils.isBlank(id) || this.chapterDao.hasChaptersBySubject(id)) continue;
+				//删除科目地区
+				this.subjectDao.deleteSubjectAreas(id);
+				//删除科目
 				this.subjectDao.deleteSubject(id);
 			}
 		}
